requirements.txt
flask
flask-cors
psycopg2-binary==2.9.9
